requirements.txt
fastapi
uvicorn
pika
bcrypt
pyjwt
psycopg2-binary
python-dotenv
redis
openai